prosopopee: save current cache when exception occurs
If an exception occurs while rendering or reencoding some files, the
whole cache is lost and whatever was done during this run of prosopopee
is lost.

Better save the work done so far so it does not need to be redone after
fixing the cause for the exception.

Note: This currently does not work for KeyboardInterrupt (i.e. Ctrl+C or
SIGINT, corrupting the Manager.dict used for the cache).

Signed-off-by: Quentin Schulz <foss@0leil.net>