cache: save cache also when Ctrl+C/SIGINT/KeyboardInterrupt is sent
The Manager server process started when calling Manager() is not able to
communicate with the main process when it's received a SIGINT making it
impossible to dump the cache in that event.

To allow graceful shutdowns, let's make the Manager server process
ignore the SIGINT signal and let the other processes raise the proper
exception so that in the try:catch the cache can be saved before
"crashing".

Shamelessly stolen from python-nose[1].

[1] https://github.com/nose-devs/nose/blob/master/nose/plugins/multiprocess.py#L144-L153

Signed-off-by: Quentin Schulz <foss@0leil.net>