prosopopee: follow progress of different long running processes
Prosopopee can take a long time to create the static content for user's
blog depending on the number of media in it.

This adds a progressbar when creating the HTML, thumbnailing pictures
and videos, and reencoding audio and video files so that the user has an
idea how much has been done and what's left in the process, alongside
the ETA for each task.

Signed-off-by: Quentin Schulz <foss@0leil.net>