update package files to follow 2021 best practices
Projects should use pyproject.toml now so that setuptools and wheels can
be used easily.

MANIFEST.in is reworked to use appropriate directives.

All content from setup.py is moved to setup.cfg which is the preferred
way today. setup.py could be removed but is still required in case one
wants to run an editable pip install of the package, so let's just leave
it be.

Signed-off-by: Quentin Schulz <foss@0leil.net>