utils: make encryption subprocess safer
By passing index_plain as stdin, there's no need to use cat and a shell
pipe to pass its content to openssl. This means, shell=True can be
removed which makes the subprocess much safer.

Also, escape the password just to be on the safe side since it is a user
setting.

This makes bandit happy.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>