tox.ini: add bandit to check for security issues
Bandit is a tool designed to find common security issues in Python code.
See https://github.com/PyCQA/bandit.

recitale has three known issues, two of low severity, related to the use
of subprocess and one of high severity which cannot currently be fixed
as it is part of a feature of recitale.

This high severity issue is related to HTML injection from settings.yaml
into the jinja2.Environment. Hopefully this can be better implemented in
the future and this security issue fixed. For now, it is explicitly
silenced.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>