add unit tests
This adds some unit tests to recitale and runs them as part of a tox
run.

A small hack was needed for reinit factories between tests so that there
is no cross-contamination which involves manually setting factory
dictionaries to an empty dict as a fixture. Hopefully something nicer
can be implemented later on.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>