recitale: video: fix call of ratio for BaseVideo
BaseVideo.size is only set by ratio property since the reencode()
function which calls ffprobe sets the size attribute only to the
ReencodedVideo and not the BaseVideo.
Thereofre, when BaseVideo.ratio is called, the size attribute does not
exist yet and will raise an AttributeError.

So instead of checking if the size attribute is set, let's check if the
object has the size attribute.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>