move thumbnail/reencode addition in own function
To prepare for the next commit, the logic possibly adding a new
thumbnail/reencode and returning it is moved to its own function. The
unit tests are adapted to reflect this change.

Technically the unit tests were a bit odd because they were testing that
the name of a pathlib.Path object of a Thumbnail/Reencode object was the
same for the addition of two different Thumbnail/Reencode objects with
the same parameters, whereas the actual desired test was to check that
two different Thumbnail/Reencode objects being added would return the
same object from the "addition function". This now allows to properly
test that.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>