handle filenames with spaces
Filenames may contain spaces (or other special characters) which are
not allowed in URLs. This is notably an issue with the exposure template
implementation for bordered-picture which uses a background-url CSS
property which only works with URLs and not filenames, therefore
filenames with spaces wouldn't work.

Let's fix that by "URL quoting" the filename so it is URL-compliant and
bordered-pictures in exposure work again.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>