use local variable for global cache
In a module, global variables are initialized at import time.
Considering that the cache module is always imported, the CACHE module
variable is always instantiated even if it's never used, i.e. in any
non-build scenario (or build with dry-run selected). By instantiating
the Cache variable only when needed, this should make recitale a tiny
bit faster to load for those cases.

Additionally, this broke recitale on systems where multiprocessing is
NOT using the fork method[1]. This is due to Python spawning a new
Python interpreter and re-importing the main module [2] (and thus the
modules it imports). This would therefore re-instantiate a Cache object
which is using Manager from multiprocessing. Manager is using another
process as server[3]. Considering that the spawning of that process
would require a reimport of the module, it results in an endless
recursive import, which triggers a RuntimeError instead of starving the
OS of its resources. By moving this into a code section that isn't
executed when importing the module, this recursive import is avoided.

[1] https://docs.python.org/3/library/multiprocessing.html#contexts-and-start-methods
[2] https://docs.python.org/3/library/multiprocessing.html#the-spawn-and-forkserver-start-methods
[3] https://docs.python.org/3/library/multiprocessing.html#sharing-state-between-processes #Server process

Fixes: #26
Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>