recitale: move local function to module namespace
For spawning start method of multiprocessing, processes are started by
spawning a new Python interpreter, importing the module and then calling
module.function (with "function" being the one passed as Pool initializer.
It is therefore necessary for this start method to have the Pool
initializer function available in the module namespace, so let's move
them there.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>