themes: material: style-page.css: make caption 100% of bordered-picture
The caption is part of a div being 100% of the media to show (be it a
video or a picture), therefore adding a right and left margin makes it
only take the centered 75% of the media instead of the full width.

The 11.5% are actually applied to the parent parent container. It's a
mistake to duplicate it, so let's remove it.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>