themes: exposure: fix video URL covers for subgalleries
In case there's a subgallery with a video as cover, the link is improper
as gallery.link provides the absolute path of the subgallery. Considering
that all paths here are expected to be relative, this ends up being
wrong. Instead, let's use the name of the subgallery instead of its link
so that the path relative to the subgallery as returned by vid and
video.thumbnail can simply be concatenated with the name to provide the
path relative to the current gallery.

Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>