autogen: include seconds in date format
TIME_FORMAT is used both for encoding and decoding a date. If a media
file doesn't contain an exif field for the date, it'll be provided one
from the mtime, itself encoded via TIME_FORMAT. However if a media file
does contain an exif field for the date, then it'll be used without
going through the encoding part. Therefore, the decoding should take
into account that the seconds part of the date can be something else
than 00.

This fixes autogen for the oldest file, in a gallery without a predefined
cover, with an exif date field where the seconds part was non-zero.

Fixes: 5e525a5d0750 ("autogen: autofill date gallery setting if missing")
Signed-off-by: Quentin Schulz <foss+recitale@0leil.net>